project: refactor with ktlint
