+++ app/data/src/main/java/com/jimmy/cryptocurrencies/data/core/Response.kt
@@ -1,4 +1,4 @@
-package com.jimmy.cryptocurrencies.common.core
+package com.jimmy.cryptocurrencies.data.core
 
 sealed class Response<out T> {
     data class Success<out T>(val data: T) : Response<T>()
