+++ app/data/src/main/java/com/jimmy/cryptocurrencies/data/local/entity/OrderBookEntity.kt
@@ -8,7 +8,8 @@ import java.util.UUID
 @Entity(tableName = "order_book_table")
 data class OrderBookEntity(
     @PrimaryKey(autoGenerate = false)
-    @ColumnInfo(name = "id") val id: String = UUID.randomUUID().toString(),
+    @ColumnInfo(name = "id")
+    val id: String = UUID.randomUUID().toString(),
     @ColumnInfo(name = "name") val name: String,
     @ColumnInfo(name = "updated_at") val updatedAt: String,
     @ColumnInfo(name = "sequence") val sequence: String
