+++ app/data/src/main/java/com/jimmy/cryptocurrencies/data/network/model/response/orderBook/AsksBidsNetworkModelResponse.kt
@@ -5,5 +5,5 @@ import com.google.gson.annotations.SerializedName
 data class AsksBidsNetworkModelResponse(
     @SerializedName("book") val book: String,
     @SerializedName("price") val price: String,
-    @SerializedName("amount") val amount: String,
+    @SerializedName("amount") val amount: String
 )
