+++ app/data/src/test/java/com/jimmy/cryptocurrencies/data/ExampleUnitTest.kt
@@ -1,7 +1,7 @@
 package com.jimmy.cryptocurrencies.data
 
-import org.junit.Test
 import org.junit.Assert.assertEquals
+import org.junit.Test
 
 /**
  * Example local unit test, which will execute on the development machine (host).
