+++ app/domain/src/main/java/com/jimmy/cryptocurrencies/domain/mapper/AvailableBookMapper.kt
