+++ app/domain/src/main/java/com/jimmy/cryptocurrencies/domain/model/availableBook/AvailableBookDomainModel.kt
