+++ app/domain/src/main/java/com/jimmy/cryptocurrencies/domain/model/orderBook/AsksBidsValueDomainModel.kt
@@ -1,6 +1,6 @@
 package com.jimmy.cryptocurrencies.domain.model.orderBook
 
-data class AsksBidsValueDomainModel (
+data class AsksBidsValueDomainModel(
     val book: String,
     val price: String,
     val amount: String
