+++ app/domain/src/main/java/com/jimmy/cryptocurrencies/domain/usecase/GetOrderBook.kt
@@ -1,7 +1,6 @@
 package com.jimmy.cryptocurrencies.domain.usecase
 
-import android.content.Context
-import com.jimmy.cryptocurrencies.common.core.Response
+import com.jimmy.cryptocurrencies.data.core.Response
 import com.jimmy.cryptocurrencies.data.repository.OrderBookRepository
 import com.jimmy.cryptocurrencies.domain.mapper.toDomainModel
 import com.jimmy.cryptocurrencies.domain.model.orderBook.OrderBookDomainModel
@@ -16,7 +15,7 @@ class GetOrderBook @Inject constructor(
             val domainResponse = repositoryResponse.toDomainModel(bookSymbol)
             Response.Success(domainResponse)
         } catch (ex: Exception) {
-            Response.Failure(code = "", message = ex.message?: "")
+            Response.Failure(code = "", message = ex.message ?: "")
         }
     }
 }
