+++ app/domain/src/test/java/com/jimmy/cryptocurrencies/domain/ExampleUnitTest.kt
@@ -1,7 +1,7 @@
 package com.jimmy.cryptocurrencies.domain
 
-import org.junit.Test
 import org.junit.Assert.assertEquals
+import org.junit.Test
 
 /**
  * Example local unit test, which will execute on the development machine (host).
