+++ app/src/test/java/com/jimmy/cryptocurrencies/ExampleUnitTest.kt
@@ -5,8 +5,8 @@ import io.mockk.clearAllMocks
 import io.mockk.every
 import io.mockk.mockk
 import org.junit.After
-import org.junit.Test
 import org.junit.Assert.assertEquals
+import org.junit.Test
 
 /**
  * Example local unit test, which will execute on the development machine (host).
